DRY code and improve ux

+++ cli/src/auth/mod.rs
@@ -2,13 +2,14 @@ use std::fs::{create_dir, read_to_string, remove_file, File};
 use std::io::prelude::*;
 use std::path::Path;
 
-use dialoguer::{console::style, theme::ColorfulTheme, Confirm};
+use dialoguer::console::style;
 use hyper::Request;
 use once_cell::sync::OnceCell;
 use serde_json::{json, Value};
 use tokio::time::{sleep, Duration};
 use webbrowser;
 
+use crate::dialoguer as dlg;
 use crate::http::CLIENT;
 
 const CODE_URL: &'static str = "https://auth.iasql.com/oauth/device/code";
@@ -38,13 +39,20 @@ pub fn get_token() -> &'static str {
 }
 
 // Get previously generated access token or generate a new one
-pub async fn authenticate(non_interactive: bool) {
+pub async fn login(non_interactive: bool) {
   let token = TOKEN.get();
   if token.is_none() {
     let home = std::env::var("HOME").unwrap();
     let file_name = &format!("{}/{}", home, TOKEN_FILE);
     match read_to_string(file_name) {
-      Ok(file_token) => TOKEN.set(file_token).unwrap(),
+      Ok(_) => {
+        let prompt = "You are already logged in. Do you wish to re-authenticate?";
+        if !dlg::confirm_with_default(&prompt, true) {
+          std::process::exit(0);
+        } else {
+          generate_token().await;
+        }
+      },
       Err(_) => match std::env::var("AUTH_TOKEN") {
         Ok(token) => TOKEN.set(token).unwrap(),
         Err(_) => {
@@ -61,10 +69,29 @@ pub async fn authenticate(non_interactive: bool) {
   };
 }
 
-pub fn clear_token() {
-  let home = std::env::var("HOME").unwrap();
-  let file_name = &format!("{}/{}", home, TOKEN_FILE);
-  remove_file(file_name).unwrap();
+pub fn logout() {
+  let token = TOKEN.get();
+  if token.is_none() {
+    let home = std::env::var("HOME").unwrap();
+    let file_name = &format!("{}/{}", home, TOKEN_FILE);
+    match read_to_string(file_name) {
+      Ok(_) => {
+        let prompt = format!("Do wish to remove the credentials stored in {}?", TOKEN_FILE);
+        if !dlg::confirm_with_default(&prompt, true) {
+          std::process::exit(0);
+        }
+        remove_file(file_name).unwrap();
+        println!(
+          "{}",
+          style("IaSQL has removed the stored credentials").bold()
+        )
+      }
+      Err(_) => println!(
+        "No stored credentials found. Call {} to generate them.",
+        style("iasql login").bold()
+      ),
+    };
+  };
 }
 
 // Prompts the user to authenticate using the Device Flow.
@@ -84,15 +111,11 @@ async fn generate_token() {
   let device_code = json["device_code"].as_str().unwrap();
   let verification_uri = json["verification_uri_complete"].as_str().unwrap();
   let user_code = json["user_code"].as_str().unwrap();
-  if !Confirm::with_theme(&ColorfulTheme::default())
-    .with_prompt(format!(
-      "{} to authenticate the IaSQL CLI in your web browser",
-      style("Press Enter").bold(),
-    ))
-    .default(true)
-    .interact()
-    .unwrap()
-  {
+  let prompt = format!(
+    "{} to authenticate the IaSQL CLI in your web browser",
+    style("Press Enter").bold(),
+  );
+  if !dlg::confirm_with_default(&prompt, true) {
     std::process::exit(0);
   }
   println!(
@@ -100,14 +123,11 @@ async fn generate_token() {
     style("!").yellow(),
     style(user_code).bold()
   );
-  let open_browser = Confirm::with_theme(&ColorfulTheme::default())
-    .with_prompt(format!(
-      "{} to open https://auth.iasql.com in your browser",
-      style("Press Enter").bold(),
-    ))
-    .default(true)
-    .interact()
-    .unwrap();
+  let prompt = format!(
+    "{} to open https://auth.iasql.com in your browser",
+    style("Press Enter").bold(),
+  );
+  let open_browser = dlg::confirm_with_default(&prompt, true);
   if !open_browser || (open_browser && webbrowser::open(verification_uri).is_err()) {
     println!(
       "Open the following url in your browser: {}",
@@ -147,15 +167,12 @@ async fn generate_token() {
       let mut file = File::create(file_name).expect(ERR);
       file.write_all(token.as_bytes()).expect(ERR);
       TOKEN.set(token.to_string()).unwrap();
-      if !Confirm::with_theme(&ColorfulTheme::default())
-        .with_prompt(format!(
-          "Authentication complete. {} to continue...",
-          style("Press Enter").bold(),
-        ))
-        .default(true)
-        .interact()
-        .unwrap()
-      {
+      let prompt = format!(
+        "Authentication complete. {} to continue...",
+        style("Press Enter").bold(),
+      );
+      if dlg::confirm_with_default(&prompt, true) {
+        println!("Welcome to IaSQL!");
         std::process::exit(0);
       }
       return;
